main loop
